test: refactor FenceConn tests to fence_driver_conn_test.go
Move all FenceConn-related tests from fence_test.go to
fence_driver_conn_test.go to improve codecov attribution.
This ensures proper test coverage reporting for fence_driver_conn.go.
